Added likes fix

+++ back-end/controllers/LikeController.js
@@ -15,7 +15,7 @@ class LikeController {
     } = req.body;
     const { user } = req; // user info from auth
     const userId = user.id;
-    const existingLike = Like.findOne({ userId, parentId });
+    const existingLike = await Like.findOne({ userId, parentId });
 
     if (existingLike) return res.status(200).json({ message: 'Already liked!' });
 
@@ -53,11 +53,8 @@ class LikeController {
 
   static async CheckUserLikedRecipe(req, res) {
     const { user } = req;
-    // console.log('User id is', user.id, req.params.recipeId);
-    const userLikes = await Like.find({
-      userId: user.id,
-      parentId: req.params.recipeId,
-    });
+    console.log('User id is', user.id, req.params.recipeId);
+    const userLikes = await Like.find({ userId: user.id, parentId: req.params.recipeId });
     return res.status(200).json(userLikes);
   }
 
@@ -77,9 +74,11 @@ class LikeController {
 
   static async DeleteLike(req, res) {
     const { user } = req;
-    const parentId = req.params.recipeId;
+    const parentId = req.body.recipeId;
 
-    const rec = await Recipe.findOne({ _id: parentId, userId: user.id });
+    const rec = await Recipe.findOne({ _id: parentId });
+    console.log(rec);
+    if (!rec) return res.status(400).json({ message: 'Recipe not found' });
     Like.deleteOne({ userId: user.id, parentId }).then(() => {
       rec.likes -= 1;
       rec.save();
@@ -90,6 +89,7 @@ class LikeController {
       console.log(err);
       res.status(400).json({ message: 'Like deletion rejected' });
     });
+    return true;
   }
 }
